+++ mmd_tools/panels/prop_lamp.py
@@ -7,11 +7,11 @@ from mmd_tools.core.lamp import MMDLamp
 
 @register_wrap
 class MMDLampPanel(Panel):
-    bl_idname = 'OBJECT_PT_mmd_tools_lamp'
-    bl_label = 'MMD Lamp Tools'
+    bl_idname = 'OBJECT_PT_mmd_tools_light'
+    bl_label = 'MMD Light Tools'
     bl_space_type = 'PROPERTIES'
     bl_region_type = 'WINDOW'
-    bl_context = 'object'
+    bl_context = 'data'
 
     @classmethod
     def poll(cls, context):
